feat: Connect BurgerIngredients and BurgerConstructor to Redux store
