--- src/contexts/ingredients-context.js
@@ -1,4 +1,0 @@
-import React from "react";
-
-
-export const IngredientsContext = React.createContext([]);
