+++ src/services/slices/burger-constructor.js
@@ -1,0 +1,18 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+
+const initialState = {
+  value: [],
+};
+
+
+export const burgerConstructorSlide = createSlice({
+  name: 'constructor',
+  initialState,
+  reducers: {
+
+  },
+});
+
+
+export default burgerConstructorSlide.reducer;
