--- src/services/slices/constructor.js
@@ -1,16 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit';
-
-
-const initialState = [];
-
-
-export const constructorSlide = createSlice({
-  name: 'constructor',
-  initialState,
-  reducers: {
-
-  },
-});
-
-
-export default constructorSlide.reducer;
